Use --error-json in runtests, distinguish simulation errors
runtests now only parses comments starting with `///`. If the first word
is `TODO`, the line is marked as a TODO. The next word indicates the
diagnostic type: `error` / `run_error` / `warning`. A lone `// TODO`
indicates something that doesn't work but should.

runtests assumes that the last line of `silq --error-json`'s output is
the JSON list of diagnostics. It may crash if the output is
syntactically valid JSON but doesn't have the right structure.